refactor(packages): moved class between packages
