+++ library/src/main/java/com/matthewtamlin/avatar/compilation/CompilerMissingException.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.avatar.element_supplier;
+package com.matthewtamlin.avatar.compilation;
 
 import javax.tools.JavaCompiler;
